+++ openoffload/cpp/framework/include/opof_serverlib.h
@@ -23,17 +23,11 @@
 #include "opof.h"
 #include "opof_error.h"
 
-
-
-//sessionTable_t * opof_create_server(const char * host, unsigned int port, const char *public_key);
-//
-//void opof_server(const char *address, unsigned short port, const char* cert, const char* key);
-
 int opof_add_session_server(sessionRequest_t *parameters, addSessionResponse_t *response);
 int opof_get_session_server(unsigned long sessionId, sessionResponse_t *response);
 int opof_del_session_server(unsigned long sessionId, sessionResponse_t *response);
 int opof_get_closed_sessions_server(statisticsRequestArgs_t *request, sessionResponse_t responses[]);
-sessionResponse_t **opof_get_all_sessions_server(statisticsRequestArgs_t *request, int *sessionCount);
+int opof_get_all_sessions_server(int pageSize, uint64_t *startSession,int pageCount, sessionResponse_t **responses);
 
 
 #endif