+++ openoffload/cpp/framework/include/opof_session_client.h
@@ -26,7 +26,6 @@
 
 extern "C" {
 #include "opof.h"
-#include "opof_error.h"
 }
 #include "opof_grpc.h"
 
@@ -34,11 +33,11 @@ class SessionTableClient {
 public: 
     SessionTableClient(std::shared_ptr<Channel> channel)
     : stub_(SessionTable::NewStub(channel)) {};
-    Status addSessionClient(int size, sessionRequest_t **s, addSessionResponse_t *resp);
-    std::string getSessionClient(int session, sessionResponse_t *resp);
-    std::string deleteSessionClient(int session, sessionResponse_t *resp);
-    void getAllSessions();
-    int getClosedSessions(statisticsRequestArgs_t *args, sessionResponse_t responses[]);
+    int addSessionClient(int size, sessionRequest_t **s, addSessionResponse_t *resp);
+    int getSessionClient(int session, sessionResponse_t *resp);
+    int deleteSessionClient(int session, sessionResponse_t *resp);
+    int getAllSessions(int pageSize, uint64_t *start_session, uint64_t *sessions, sessionResponse_t responses[],unsigned long *sessionCount);
+    int getClosedSessions(statisticsRequestArgs_t *args, sessionResponse_t responses[], unsigned long *sessionCount);
 private:
     std::unique_ptr<SessionTable::Stub> stub_;
 };