+++ openoffload/cpp/framework/src/opof_session_server.cc
@@ -41,31 +41,28 @@ extern "C" {
 * \param response
 */
 Status SessionTableImpl::addSession(ServerContext* context, ServerReader<sessionRequest>* reader, addSessionResponse* response) {
-    int status;
-    unsigned long errorCode;
-    unsigned long error = 1;
-    int index=0;
-    ADD_SESSION_STATUS reqStatus;
-    addSessionResponse_t addResponse_c;
-    sessionRequest_t request_c;
-    sessionRequest request;
-    while(reader->Read(&request)){
-      convertSessionRequest2c(request, &request_c);
-      status = opof_add_session_server(&request_c, &addResponse_c);
-      if (status == FAILURE){
-        errorCode = errorCode ^ (error << index);
-      }
-      index++;
+  int status;
+  //int index=0;
+  //ADD_SESSION_STATUS reqStatus = ADD_SESSION_STATUS::_SESSION_ACCEPTED;
+  addSessionResponse_t addResponse_c;
+  sessionResponseError *errorMessage;
+  sessionRequest_t request_c;
+  sessionRequest request;
+  response->clear_responseerror();
+  //std::cout << "response size: " << response->responseerror_size() << std::endl;
+  while(reader->Read(&request)){
+    convertSessionRequest2c(request, &request_c);
+    status = opof_add_session_server(&request_c, &addResponse_c);
+    if (status != _OK){
+      errorMessage = response->add_responseerror();
+      errorMessage->set_sessionid(request.sessionid());
+      errorMessage->set_errorstatus(status);
     }
-    if (errorCode > 0){
-      reqStatus = ADD_SESSION_STATUS::_SESSION_REJECTED;
-    }else {
-      reqStatus = ADD_SESSION_STATUS::_SESSION_ACCEPTED;
-    }
-    response->set_requeststatus(reqStatus);
-    response->set_errorstatus(errorCode);
-    return Status::OK;
+    //index++;
   }
+  //response->set_requeststatus(reqStatus);
+  return Status::OK;
+}
 
 /** \ingroup serverlibrary
 * \brief getSession
@@ -75,18 +72,21 @@ Status SessionTableImpl::addSession(ServerContext* context, ServerReader<session
 * \param response
 */
 Status SessionTableImpl::getSession(ServerContext* context, const sessionId* sid,
-                  sessionResponse* response) {
-        sessionResponse_t response_c;
-        int status;
-
-        status = opof_get_session_server(sid->sessionid(), &response_c);
-        if (status == SUCCESS){
-          convertSessionResponse2cpp(response, &response_c);
-          return Status::OK;
-        } else {
-          return Status(grpc::StatusCode::NOT_FOUND, "Get Session Not Found");
-        }
+  sessionResponse* response) {
+  sessionResponse_t response_c;
+  int status;
+  uint64_t session;
+  session = sid->sessionid();
+  status = opof_get_session_server(session, &response_c);
+  if (status == _OK){
+    convertSessionResponse2cpp(response, &response_c);
+    return Status::OK;
+  } else if (status == _NOT_FOUND){
+    return Status(grpc::StatusCode::NOT_FOUND, "Get Session Not Found");
+  } else {
+    return Status(grpc::StatusCode::UNKNOWN, "Internal Open Offload Server Error");
   }
+}
 
 /** \ingroup serverlibrary
 * \brief deleteSession
@@ -96,18 +96,19 @@ Status SessionTableImpl::getSession(ServerContext* context, const sessionId* sid
 * \param response
 */
 Status SessionTableImpl::deleteSession(ServerContext* context, const sessionId* sid,
-                  sessionResponse* response) {
-      int status;
-      sessionResponse_t response_c;
-        //std::cout << "sessionID to delete is: " << sid->sessionid() << std::endl;
-        status = opof_del_session_server(sid->sessionid(), &response_c);
-        if (status ==SUCCESS){
-          convertSessionResponse2cpp(response, &response_c);
-          return Status::OK;
-        } else {
-          return Status(grpc::StatusCode::NOT_FOUND, "Delete Session Not Found");
-        }
+  sessionResponse* response) {
+  int status;
+  sessionResponse_t response_c;
+  status = opof_del_session_server(sid->sessionid(), &response_c);
+  if (status == _OK){
+    convertSessionResponse2cpp(response, &response_c);
+    return Status::OK;
+  } else if (status == _NOT_FOUND){
+    return Status(grpc::StatusCode::NOT_FOUND, "Delete Session Not Found");
+  } else {
+    return Status(grpc::StatusCode::UNKNOWN, "Internal Open Offload Server Error");
   }
+}
 
 /** \ingroup serverlibrary
 * \brief getAllSessions
@@ -116,37 +117,49 @@ Status SessionTableImpl::deleteSession(ServerContext* context, const sessionId* 
 * \param reader
 * \param response
 */
-  Status SessionTableImpl::getAllSessions(ServerContext* context, const statisticsRequestArgs* request, ServerWriter<sessionResponse>* writer) {
+Status SessionTableImpl::getAllSessions(ServerContext* context, const statisticsRequestArgs* request, sessionResponseArray *responses) {
   
-  sessionResponse response;
-  sessionResponse_t **closedSessions= NULL;
+  Status status;
+  sessionResponse_t **allSessions= NULL;
   sessionResponse_t *closedResponse;
-  statisticsRequestArgs_t request_c;
-  int sessionCount = 0;
-  int nresponses = request->pagesize();
-  request_c.pageSize = nresponses;
-  closedSessions = opof_get_all_sessions_server(&request_c, &sessionCount);
-  if (closedSessions == NULL){
-    response.set_requeststatus(REQUEST_STATUS::_NO_CLOSED_SESSIONS);
-    return Status::OK;
+  sessionResponse *response;
+  int sessionCount;
+  int pageCount = 0;
+  
+  int nresponses = BUFFER_MAX;
+  uint64_t start_session;
+  start_session = request->startsession();
+  
+  allSessions = (sessionResponse_t **)malloc(nresponses * sizeof(sessionResponse_t *));
+  for (int i = 0; i < nresponses; i++){
+    allSessions[i] = (sessionResponse_t *)malloc(sizeof(sessionResponse_t));
   }
-  for (int i=0; i < sessionCount; i++){
-    closedResponse = closedSessions[i];
-    response.set_sessionid(closedResponse->sessionId);
-    response.set_sessionstate((SESSION_STATE)closedResponse->sessionState);
-    response.set_inpackets(closedResponse->inPackets);
-    response.set_outpackets(closedResponse->outPackets);
-    response.set_inbytes(closedResponse->inBytes);
-    response.set_outbytes(closedResponse->outBytes);
-    response.set_sessionclosecode((SESSION_CLOSE_CODE)closedResponse->sessionCloseCode);
-    response.set_requeststatus(REQUEST_STATUS::_ACCEPTED);
-    writer->Write(response);
-    free(closedResponse);
-  }
-  free(closedSessions);
+  
+  sessionCount = opof_get_all_sessions_server(nresponses, &start_session, pageCount, allSessions);
 
+  responses->set_nextkey(start_session);
+  
+  if (sessionCount > 0){
+    pageCount++;
+    for (int i=0; i < sessionCount; i++){
+      closedResponse = allSessions[i];
+      response = responses->add_responsearray();
+      response->set_sessionid(closedResponse->sessionId);
+      response->set_sessionstate((SESSION_STATE)closedResponse->sessionState);
+      response->set_inpackets(closedResponse->inPackets);
+      response->set_outpackets(closedResponse->outPackets);
+      response->set_inbytes(closedResponse->inBytes);
+      response->set_outbytes(closedResponse->outBytes);
+      response->set_sessionclosecode((SESSION_CLOSE_CODE)closedResponse->sessionCloseCode);
+      response->set_requeststatus((REQUEST_STATUS)closedResponse->requestStatus);
+    }
+  }
+  for (int i=0; i <nresponses; i++){
+    free(allSessions[i]);
+  }
+  free(allSessions);
   return Status::OK;
- }
+}
 
 /** \ingroup serverlibrary
 * \brief getClosedSessions
@@ -161,12 +174,15 @@ Status SessionTableImpl::getClosedSessions(ServerContext* context, const statist
   sessionResponse_t closedResponse;
   statisticsRequestArgs_t request_c;
   int sessionCount = 0;
+  //Status status;
   int nresponses = request->pagesize();
   request_c.pageSize = nresponses;
   sessionResponse_t closedSessions[BUFFER_MAX];
 
   sessionCount = opof_get_closed_sessions_server(&request_c, closedSessions);
-
+  if (sessionCount == 0){
+    return Status(grpc::StatusCode::NOT_FOUND,"No Closed Sessions");
+  }
   for (int i=0; i < sessionCount; i++){
     closedResponse = closedSessions[i];
     response.set_sessionid(closedResponse.sessionId);
@@ -182,4 +198,4 @@ Status SessionTableImpl::getClosedSessions(ServerContext* context, const statist
   }
   
   return Status::OK;
- }
+}