+++ openoffload/cpp/framework/src/opof_util.cc
@@ -31,29 +31,6 @@ extern "C" {
 #include "opof_util.h"
 #include "opof_grpc.h"
 
-
-
-/** \ingroup utilities
-*
-* \brief Append the port to the address string and return as string.
-*
-* \param *address   A string containing the address of the server. The string must
-*                   be long enough to append the port too.
-* \param prot       The listening port as an int, it will be converted to a string
-*                   and appended to the address string.
-*
-* \return SUCCESS or FAILURE
-*/
-int create_address(char *address, unsigned short port){
-  int status = SUCCESS;
-  char port_str[12];
-  sprintf(port_str, "%u", port);
-  strcat(address,":");
-  strcat(address,port_str);
-  printf("Address: %s\n",address);
-
-  return status;
-}
 /** \ingroup utilities
 *
 * \brief Covert a C SessionRequest_t to a C++ sessionRequest Class instance
@@ -71,7 +48,7 @@ void convertSessionRequest2cpp(sessionRequest_t *request_c, sessionRequest *requ
     request->set_inlif(request_c->inlif);
     request->set_outlif(request_c->outlif);
     request->set_ipversion((IP_VERSION)request_c->ipver);
-    request->set_sourceport(request_c->srcPort);
+    request->set_sourceport((unsigned int)request_c->srcPort);
     if (request_c->ipver == _IPV6){
        s.assign(request_c->srcIPV6.s6_addr, request_c->srcIPV6.s6_addr+ 16);
       request->set_sourceipv6(s);
@@ -84,7 +61,7 @@ void convertSessionRequest2cpp(sessionRequest_t *request_c, sessionRequest *requ
     } else {
       request->set_destinationip(request_c->dstIP.s_addr);
     } 
-    request->set_destinationport(request_c->dstPort);
+    request->set_destinationport((unsigned int)request_c->dstPort);
     request->set_protocolid((PROTOCOL_ID)request_c->proto);
     action.set_actiontype((ACTION_TYPE)request_c->actType);
     if (request_c->ipver == _IPV6){
@@ -94,6 +71,7 @@ void convertSessionRequest2cpp(sessionRequest_t *request_c, sessionRequest *requ
       action.set_actionnexthop(request_c->nextHop.s_addr);
     }
     request->mutable_action()->CopyFrom(action);
+    request->set_cachetimeout(request_c->cacheTimeout);
 
 }
 /** \ingroup utilities
@@ -107,9 +85,18 @@ void convertSessionRequest2cpp(sessionRequest_t *request_c, sessionRequest *requ
 * \return void
 */
 void convertAddSessionResponse2c(addSessionResponse_t *response_c, addSessionResponse *response){
-
-  response_c->requestStatus = (REQUEST_STATUS_T)response->requeststatus();
-  response_c->errorStatus = response->errorstatus();
+  sessionResponseError responseError;
+  //response_c->requestStatus = (REQUEST_STATUS_T)response->requeststatus();
+  if (response->responseerror_size() > 0){
+    response_c->number_errors = response->responseerror_size();
+    for (int i=0; i< response_c->number_errors; i++){
+      responseError = response->responseerror(i);
+      response_c->sessionErrors[i].sessionId = responseError.sessionid();
+      response_c->sessionErrors[i].errorStatus = responseError.errorstatus();
+    }
+  } else {
+    response_c->number_errors = 0;
+  }
 }
 /** \ingroup utilities
 *
@@ -162,12 +149,12 @@ void convertSessionRequest2c(sessionRequest request, sessionRequest_t *request_c
     } else {
       request_c->dstIP.s_addr = request.destinationip();
     }
-    request_c->srcPort = request.sourceport();
-    request_c->dstPort = request.destinationport();
+    request_c->srcPort = (unsigned short)request.sourceport();
+    request_c->dstPort = (unsigned short)request.destinationport();
     request_c->proto = (PROTOCOL_ID_T)request.protocolid();
     action = request.action();
     request_c->actType= (ACTION_VALUE_T)action.actiontype();
-
+    request_c->cacheTimeout = request.cachetimeout();
      if (request_c->ipver == _IPV6){
       char_array.assign(action.actionnexthopv6().begin(), action.actionnexthopv6().end());
       memcpy(request_c->nextHopV6.s6_addr,&char_array[0],16);