+++ tests/basic/sessions_server.py
@@ -26,8 +26,42 @@ import sys
 import grpc
 import google.protobuf.timestamp_pb2
 
+#from grpc_status import rpc_status
+
+from google.protobuf import any_pb2
+#from google.rpc import code_pb2, status_pb2, error_details_pb2
+
 import openoffload_pb2
 import openoffload_pb2_grpc
+
+
+
+class AddSessionErrors:
+   '''
+   '''
+   def __init__(self):
+       self._addSessionErrors= list()
+   def addSessionErrorMembers(self, sessionResponseError):
+       self._addSessionErrors.append(sessionResponseError)
+   def __iter__(self):
+       ''' Returns the Iterator object '''
+       return AddSessionErrorsIterator(self)
+
+class AddSessionErrorsIterator:
+   ''' Iterator class '''
+   def __init__(self, addSessionErrors_list):
+       # AddSessionErrors object reference
+       self._addSessionErrors_list = addSessionErrors_list
+       # member variable to keep track of current index
+       self._index = 0
+   def __next__(self):
+       ''''Returns the next value from team object's lists '''
+       if self._index < (len(self._addSessionErrors_list._addSessionErrors) ) :
+           result = (self._addSessionErrors_list._addSessionErrors[self._index])
+           self._index +=1
+           return result
+       # End of Iteration
+       raise StopIteration
 
 
 class SessionTableServicer(openoffload_pb2_grpc.SessionTableServicer):
@@ -43,65 +77,78 @@ class SessionTableServicer(openoffload_pb2_grpc.SessionTableServicer):
         """do some init stuff"""
 
     def addSession(self, request_iterator, context):
+        sessionErrors_value=AddSessionErrors()
         for request in request_iterator:
-            print("############ ADD SESSION ##################");
-            #print("Application SessionID:", request.sessionId);
-            print("protocolID 6=TCP,17=UDP:",request.protocolId);
+            print("############ ADD SESSION ##################")
+            print("protocolID 6=TCP,17=UDP:",request.protocolId)
             print("IP Version:", openoffload_pb2._IP_VERSION.values_by_number[request.ipVersion].name)
+            
 
             if request.ipVersion == openoffload_pb2._IPV4:
               print ("sourceIp:", socket.inet_ntop(socket.AF_INET, request.sourceIp.to_bytes(4,byteorder=sys.byteorder)))
             else:
               print ("sourceIpV6:", socket.inet_ntop(socket.AF_INET6, request.sourceIpV6))
-            print("sourcePort:", int(request.sourcePort));
+            print("sourcePort:", int(request.sourcePort))
             if request.ipVersion == openoffload_pb2._IPV4:
               print ("destinationIp:", socket.inet_ntop(socket.AF_INET, request.destinationIp.to_bytes(4,byteorder=sys.byteorder)))
             else:
               print ("destinationIpV6:", socket.inet_ntop(socket.AF_INET6, request.destinationIpV6))
-            print("destinationPort:", int(request.destinationPort));
+            print("destinationPort:", int(request.destinationPort))
             print("ActionType 0=DROP,1=FORWARD,2=MIRROR,3=SNOOP:" , request.action.actionType)
             print("ActionNextHop:" , request.action.actionNextHop)
-        return openoffload_pb2.addSessionResponse(requestStatus=openoffload_pb2._ACCEPTED);
+            if  request.sessionId == 99999999999:
+                print("Error test case")
+                #sessionError=openoffload_pb2.addSessionResponse.responseError()
+                sessionError=openoffload_pb2.sessionResponseError()
+                sessionError.sessionId=request.sessionId
+                sessionError.errorStatus=openoffload_pb2._SESSION_TABLE_FULL
+                sessionErrors_value.addSessionErrorMembers( sessionError)
+
+        return openoffload_pb2.addSessionResponse(requestStatus=openoffload_pb2._ACCEPTED, responseError=sessionErrors_value)
 
     def getSession(self, request, context):
-            print("############ GET SESSION ##################");
-            print("sessionId:",request.sessionId);
+            print("############ GET SESSION ##################")
+            print("sessionId:",request.sessionId)
             timestamp = google.protobuf.timestamp_pb2.Timestamp()
             timestamp.GetCurrentTime()
             return openoffload_pb2.sessionResponse(sessionId=1001,sessionState=openoffload_pb2._ESTABLISHED,
               requestStatus=openoffload_pb2._ACCEPTED, inPackets=1000, outPackets=200000,
-              startTime=timestamp);
+              startTime=timestamp)
     def deleteSession(self, request, context):
-            print("############ DELETE SESSION ##################");
-            print("sessionId:",request.sessionId);
+            print("############ DELETE SESSION ##################")
+            print("sessionId:",request.sessionId)
             timestamp = google.protobuf.timestamp_pb2.Timestamp()
             timestamp.GetCurrentTime()
             return openoffload_pb2.sessionResponse(sessionId=1001, sessionState=openoffload_pb2._CLOSING_1,
               requestStatus=openoffload_pb2._ACCEPTED, inPackets=2000, outPackets=400000,
-              startTime=timestamp, endTime=timestamp);
+              startTime=timestamp, endTime=timestamp)
 
     def getClosedSessions(self, request, context):
-            print("############ GET CLOSED SESSIONS ##################");
+            print("############ GET CLOSED SESSIONS ##################")
             timestamp = google.protobuf.timestamp_pb2.Timestamp()
             timestamp.GetCurrentTime()
             session1 = openoffload_pb2.sessionResponse(sessionId=1001, sessionState=openoffload_pb2._CLOSED,
-              requestStatus=openoffload_pb2._ACCEPTED, inPackets=1000, outPackets=200000, startTime=timestamp);
+              requestStatus=openoffload_pb2._ACCEPTED, inPackets=1000, outPackets=200000, startTime=timestamp)
             session2 = openoffload_pb2.sessionResponse(sessionId=1002, sessionState=openoffload_pb2._CLOSED,
-              requestStatus=openoffload_pb2._ACCEPTED, inPackets=2000, outPackets=400000, startTime=timestamp);
+              requestStatus=openoffload_pb2._ACCEPTED, inPackets=2000, outPackets=400000, startTime=timestamp)
             yield session1
             yield session2
 
     def getAllSessions(self, request, context):
-            print("############ GET ALL SESSIONS ##################");
+            print("############ GET ALL SESSIONS ##################")
+            sessionResponseArray = openoffload_pb2.sessionResponseArray()
             timestamp = google.protobuf.timestamp_pb2.Timestamp()
             timestamp.GetCurrentTime()
-            session1 = openoffload_pb2.sessionResponse(sessionId=1001, sessionState=openoffload_pb2._CLOSED,
-              requestStatus=openoffload_pb2._ACCEPTED, inPackets=1000, outPackets=200000, startTime=timestamp);
-            session2 = openoffload_pb2.sessionResponse(sessionId=1002, sessionState=openoffload_pb2._CLOSED,
-              requestStatus=openoffload_pb2._ACCEPTED, inPackets=2000, outPackets=400000, startTime=timestamp);
-            yield session1
-            yield session2
 
+            session1 = openoffload_pb2.sessionResponse(sessionId=1001, sessionState=openoffload_pb2._CLOSED,
+              requestStatus=openoffload_pb2._ACCEPTED, inPackets=1000, outPackets=200000, startTime=timestamp)
+            sessionResponseArray.responseArray.append(session1)
+
+            session2 = openoffload_pb2.sessionResponse(sessionId=1002, sessionState=openoffload_pb2._CLOSED,
+              requestStatus=openoffload_pb2._ACCEPTED, inPackets=2000, outPackets=400000, startTime=timestamp)
+            sessionResponseArray.responseArray.append(session2)
+
+            return sessionResponseArray
 
 
 def sessionServe():